connect_safe funtion
